+++ bot/src/main/java/edu/java/bot/command/Command.java
@@ -1,4 +1,4 @@
-package edu.java.bot;
+package edu.java.bot.command;
 
 import com.pengrad.telegrambot.model.BotCommand;
 import com.pengrad.telegrambot.model.Update;
