+++ bot/src/main/java/edu/java/bot/command/StartCommand.java
@@ -1,4 +1,4 @@
-package edu.java.bot;
+package edu.java.bot.command;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.model.request.ParseMode;
