+++ bot/src/main/java/edu/java/bot/command/UntrackCommand.java
@@ -1,4 +1,4 @@
-package edu.java.bot;
+package edu.java.bot.command;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
