+++ bot/src/test/java/edu/java/bot/command/ListCommandTest.java
@@ -3,7 +3,6 @@ package edu.java.bot.command;
 import com.pengrad.telegrambot.model.Chat;
 import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.Update;
-import edu.java.bot.ListCommand;
 import edu.java.bot.model.Link;
 import edu.java.bot.model.User;
 import edu.java.bot.service.UserService;
