+++ bot/src/test/java/edu/java/bot/command/StartCommandTest.java
@@ -3,7 +3,6 @@ package edu.java.bot.command;
 import com.pengrad.telegrambot.model.Chat;
 import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.Update;
-import edu.java.bot.StartCommand;
 import edu.java.bot.model.User;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
